Created block subpackage in several packages related to animation, renamed RotatingHeadAnimatedModel(was AnimatedModel) for more clarity of use
