+++ src/main/java/com/mraof/minestuck/client/model/ConsortModel.java
@@ -4,7 +4,7 @@ import com.mraof.minestuck.Minestuck;
 import com.mraof.minestuck.entity.consort.ConsortEntity;
 import net.minecraft.resources.ResourceLocation;
 
-public class ConsortModel<T extends ConsortEntity> extends AnimatedModel<T>
+public class ConsortModel<T extends ConsortEntity> extends RotatingHeadAnimatedModel<T>
 {
 	@Override
 	public ResourceLocation getModelLocation(T entity)
