+++ src/main/java/com/mraof/minestuck/client/model/PawnModel.java
@@ -5,7 +5,7 @@ import com.mraof.minestuck.entity.carapacian.EnumEntityKingdom;
 import com.mraof.minestuck.entity.carapacian.PawnEntity;
 import net.minecraft.resources.ResourceLocation;
 
-public class PawnModel extends AnimatedModel<PawnEntity>
+public class PawnModel extends RotatingHeadAnimatedModel<PawnEntity>
 {
 	@Override
 	public ResourceLocation getModelLocation(PawnEntity object)
