+++ src/main/java/com/mraof/minestuck/client/model/RotatingHeadAnimatedModel.java
@@ -11,7 +11,7 @@ import javax.annotation.Nullable;
 /**
  * Base class for all entity models which implements head movement based on the look direction
  */
-public abstract class AnimatedModel<T extends IAnimatable> extends AnimatedGeoModel<T> {
+public abstract class RotatingHeadAnimatedModel<T extends IAnimatable> extends AnimatedGeoModel<T> {
 	@Override
 	public void setLivingAnimations(T entity, Integer uniqueID, @Nullable AnimationEvent predicate) {
 		super.setLivingAnimations(entity, uniqueID, predicate);
