+++ src/main/java/com/mraof/minestuck/client/renderer/blockentity/AlchemiterRenderer.java
@@ -1,6 +1,6 @@
 package com.mraof.minestuck.client.renderer.blockentity;
 
-import com.mraof.minestuck.client.model.machine.AlchemiterModel;
+import com.mraof.minestuck.client.model.block.AlchemiterModel;
 import com.mraof.minestuck.blockentity.machine.AlchemiterBlockEntity;
 import net.minecraft.client.renderer.blockentity.BlockEntityRendererProvider;
 import software.bernie.geckolib3.renderers.geo.GeoBlockRenderer;
