+++ src/main/java/com/mraof/minestuck/client/renderer/blockentity/TotemLatheRenderer.java
@@ -7,9 +7,8 @@ import com.mraof.minestuck.alchemy.AlchemyHelper;
 import com.mraof.minestuck.block.CruxiteDowelBlock;
 import com.mraof.minestuck.block.MSBlocks;
 import com.mraof.minestuck.block.MSProperties;
-import com.mraof.minestuck.blockentity.machine.TotemLatheBlockEntity;
 import com.mraof.minestuck.blockentity.machine.TotemLatheDowelBlockEntity;
-import com.mraof.minestuck.client.model.TotemLatheModel;
+import com.mraof.minestuck.client.model.block.TotemLatheModel;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.renderer.ItemBlockRenderTypes;
