+++ src/main/resources/assets/minestuck/animations/block/alchemiter.animation.json
