+++ src/main/resources/assets/minestuck/animations/block/totemlathe.animation.json
