+++ src/main/resources/assets/minestuck/geo/block/alchemiter.geo.json
