+++ src/main/resources/assets/minestuck/geo/block/totemlathe.geo.json
